+++ packages/store/src/decorators/selector.ts
@@ -1,57 +1,14 @@
-import { memoize } from '../utils/memoize';
-import { getValue } from '../utils/utils';
-import { ensureStoreMetadata, ensureSelectorMetadata } from '../internal/internals';
-import { getSelectorFn } from '../utils/selector-utils';
+import { createSelector } from '../utils/selector-utils';
 
 /**
  * Decorator for memoizing a state selector.
  */
 export function Selector(selectors?: any[]) {
   return (target: any, key: string, descriptor: PropertyDescriptor) => {
-    const metadata = ensureStoreMetadata(target);
-
     if (descriptor.value !== null) {
-      const prev = descriptor.value;
-      const wrappedFn = function wrappedSelectorFn(...args) {
-        const returnValue = prev(...args);
-        if (returnValue instanceof Function) {
-          const innerMemoizedFn = memoize.apply(null, [returnValue]);
-          return innerMemoizedFn;
-        }
-        return returnValue;
-      };
-      const memoizedFn = memoize(wrappedFn);
+      const originalFn = descriptor.value;
 
-      const fn = state => {
-        const results = [];
-
-        // If we are on a state class, get the metadata path
-        if (metadata && metadata.path) {
-          results.push(getValue(state, metadata.path));
-        }
-
-        // Allow additional selectors if passed
-        if (selectors) {
-          results.push(...selectors.map(a => getSelectorFn(a)(state)));
-        }
-
-        // if the lambda tries to access a something on the
-        // state that doesn't exist, it will throw a TypeError.
-        // since this is quite usual behaviour, we simply return undefined if so.
-        try {
-          return memoizedFn(...results);
-        } catch (ex) {
-          if (ex instanceof TypeError) {
-            return undefined;
-          }
-          throw ex;
-        }
-      };
-
-      const selectorMetaData = ensureSelectorMetadata(memoizedFn);
-      selectorMetaData.originalFn = prev;
-      selectorMetaData.storeMetaData = metadata;
-      selectorMetaData.selectFromAppState = fn;
+      const memoizedFn = createSelector(selectors, originalFn, { containerClass: target, selectorName: key });
 
       return {
         configurable: true,